Refactor track classes.
1. Abstract common code of AudioTrack and VideoTrack into TrackBase.
2. Add AudioTrackList::HasEnabledTrack().
3. Raise correct exception in TrackDefault and TrackDefaultList
   implementations.
4. Fix return type of VideoTrackList::selected_index() as its return
   value can be -1.

b/35259101

+++ cobalt/dom/video_track_list.h
@@ -41,11 +41,11 @@ class VideoTrackList : public TrackListBase<VideoTrack> {
   using TrackListBase<VideoTrack>::AnonymousIndexedGetter;
   using TrackListBase<VideoTrack>::GetTrackById;
 
-  uint32 selected_index() const {
+  int32 selected_index() const {
     for (uint32 i = 0; i < length(); ++i) {
       scoped_refptr<VideoTrack> video_track = AnonymousIndexedGetter(i);
       if (video_track->selected()) {
-        return i;
+        return static_cast<int32>(i);
       }
     }
